Refactor file stat and ensure Linux gets correct ctime
Look for birthtime if supported on systems. Also, move file attribute
functions to util.
