+++ rummage/lib/gui/data/docs/.dochash
@@ -1,1 +1,1 @@
-faa2ffe1ce652963b312fbd75e6d4f21
+28779be569f92b6d78c94b5e98470719
