+++ rummage/lib/gui/data/docs/changelog.html
@@ -25,6 +25,10 @@
 
   <!-- Table of Content (Don't show for static sites) -->
     <h1 id="changelog">Changelog</h1>
+<h2 id="current">Current</h2>
+<ul>
+<li><strong>FIX</strong>: Better attempt to get appropriate creation on all Linux systems.</li>
+</ul>
 <h2 id="463">4.6.3</h2>
 <ul>
 <li><strong>FIX</strong>: Notification audio failures should be handled gracefully.</li>
